split ServerSettings::Database class
